requirements.txt
ytmusicapi~=0.25
pandas~=1.4
python-dotenv~=1.0